Move copy/paste related terminal logic to clipboard terminalContrib/
Part of #230129
